+++ src/main/java/org/entur/lamassu/controller/GBFSInternalFeedController.java
@@ -1,0 +1,179 @@
+/*
+ *
+ *
+ *  * Licensed under the EUPL, Version 1.2 or – as soon they will be approved by
+ *  * the European Commission - subsequent versions of the EUPL (the "Licence");
+ *  * You may not use this work except in compliance with the Licence.
+ *  * You may obtain a copy of the Licence at:
+ *  *
+ *  *   https://joinup.ec.europa.eu/software/page/eupl
+ *  *
+ *  * Unless required by applicable law or agreed to in writing, software
+ *  * distributed under the Licence is distributed on an "AS IS" basis,
+ *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  * See the Licence for the specific language governing permissions and
+ *  * limitations under the Licence.
+ *
+ */
+
+package org.entur.lamassu.controller;
+
+import java.net.URI;
+import java.time.Instant;
+import java.util.NoSuchElementException;
+import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
+import org.entur.gbfs.v2_3.gbfs.GBFS;
+import org.entur.gbfs.v2_3.gbfs.GBFSFeed;
+import org.entur.gbfs.v2_3.gbfs.GBFSFeedName;
+import org.entur.gbfs.v2_3.gbfs.GBFSFeeds;
+import org.entur.lamassu.cache.GBFSFeedCache;
+import org.entur.lamassu.model.discovery.System;
+import org.entur.lamassu.model.discovery.SystemDiscovery;
+import org.entur.lamassu.model.provider.FeedProvider;
+import org.entur.lamassu.service.FeedProviderService;
+import org.entur.lamassu.service.SystemDiscoveryService;
+import org.entur.lamassu.util.CacheUtil;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.CacheControl;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
+
+@RestController
+public class GBFSInternalFeedController extends BaseGBFSFeedController {
+
+  private final SystemDiscoveryService systemDiscoveryService;
+
+  private final FeedProviderService feedProviderService;
+
+  @Value("${org.entur.lamassu.baseUrl}")
+  private String baseUrl;
+
+  @Value("${org.entur.lamassu.internalLoadBalancer}")
+  private String internalLoadBalancer;
+
+  @Autowired
+  public GBFSInternalFeedController(
+    SystemDiscoveryService systemDiscoveryService,
+    GBFSFeedCache feedCache,
+    FeedProviderService feedProviderService
+  ) {
+    super(feedCache, feedProviderService);
+    this.systemDiscoveryService = systemDiscoveryService;
+    this.feedProviderService = feedProviderService;
+  }
+
+  @GetMapping("/gbfs-internal")
+  public ResponseEntity<SystemDiscovery> getInternalFeedProviderDiscovery() {
+    var data = systemDiscoveryService.getSystemDiscovery();
+    return ResponseEntity
+      .ok()
+      .cacheControl(CacheControl.maxAge(60, TimeUnit.MINUTES).cachePublic())
+      .body(modifySystemDiscoveryUrls(data));
+  }
+
+  @GetMapping(
+    value = {
+      "/gbfs-internal/{systemId}/{feed}", "/gbfs-internal/{systemId}/{feed}.json",
+    }
+  )
+  public Object getInternalGbfsFeedForProvider(
+    @PathVariable String systemId,
+    @PathVariable String feed
+  ) {
+    try {
+      var feedName = GBFSFeedName.fromValue(feed);
+      var feedProvider = feedProviderService.getFeedProviderBySystemId(systemId);
+      Object data = getFeed(systemId, feed);
+      if (feedName.equals(GBFSFeedName.GBFS)) {
+        data = modifyDiscoveryUrls(feedProvider, (GBFS) data);
+      }
+
+      return ResponseEntity
+        .ok()
+        .cacheControl(
+          CacheControl
+            .maxAge(
+              CacheUtil.getMaxAge(
+                feedName.implementingClass(),
+                data,
+                systemId,
+                feed,
+                (int) Instant.now().getEpochSecond()
+              ),
+              TimeUnit.SECONDS
+            )
+            .cachePublic()
+        )
+        .lastModified(
+          CacheUtil.getLastModified(feedName.implementingClass(), data, systemId, feed)
+        )
+        .body(data);
+    } catch (IllegalArgumentException e) {
+      throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+    } catch (NoSuchElementException e) {
+      throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+    }
+  }
+
+  private SystemDiscovery modifySystemDiscoveryUrls(SystemDiscovery data) {
+    var systemDiscovery = new SystemDiscovery();
+    systemDiscovery.setSystems(
+      data
+        .getSystems()
+        .stream()
+        .map(s -> {
+          var system = new System();
+          system.setId(s.getId());
+          system.setUrl(
+            s.getUrl().replace(baseUrl + "/gbfs", internalLoadBalancer + "/gbfs-internal")
+          );
+          return system;
+        })
+        .collect(Collectors.toList())
+    );
+    return systemDiscovery;
+  }
+
+  private GBFS modifyDiscoveryUrls(FeedProvider feedProvider, GBFS data) {
+    var gbfs = new GBFS();
+    gbfs.setLastUpdated(data.getLastUpdated());
+    gbfs.setTtl(data.getTtl());
+    gbfs.setFeedsData(
+      data
+        .getFeedsData()
+        .values()
+        .stream()
+        .map(gbfsFeeds -> {
+          var mappedGbfsFeeds = new GBFSFeeds();
+          mappedGbfsFeeds.setFeeds(
+            gbfsFeeds
+              .getFeeds()
+              .stream()
+              .map(f -> {
+                var gbfsFeed = new GBFSFeed();
+                gbfsFeed.setName(f.getName());
+                gbfsFeed.setUrl(
+                  URI.create(
+                    f
+                      .getUrl()
+                      .toString()
+                      .replace(baseUrl + "/gbfs", internalLoadBalancer + "/gbfs-internal")
+                  )
+                );
+                return gbfsFeed;
+              })
+              .collect(Collectors.toList())
+          );
+          return mappedGbfsFeeds;
+        })
+        .collect(Collectors.toMap(e -> feedProvider.getLanguage(), e -> e))
+    );
+    return gbfs;
+  }
+}
